+++ tests/common/MockLocalStorage.swift
@@ -41,7 +41,9 @@ class MockLocalStorage: LocalStorageProtocol {
     var isNotificationsEnabled: Bool = false
     
     var hasStoredNotificationSetting: Bool = false
-    
+
+    var keychainMigrationCompleted: Bool = false
+
     func getAttributionInfo(currentDate: Date) -> IterableAttributionInfo? {
         guard !MockLocalStorage.isExpired(expiration: attributionInfoExpiration, currentDate: currentDate) else {
             return nil